+++ packages/core/src/components/ozon-content/ozon-content-node.interface.ts
@@ -9,5 +9,5 @@ export interface OzonContentNode<T = unknown> {
 
   identify?(node: Node): string | undefined;
 
-  render(node: Node, context: OzonContentNodeContext<T>): JSX.Element;
+  render(node: Node, context: OzonContentNodeContext<T>, ignoreMark?: boolean): JSX.Element;
 }